[WIP] Attempted to implement something for mainpulating inventory via a "showlist" command
It would also have a counter-part "setlist". There was some attempt to create some capability to
walk/traverse a datastructure deeply. However, working through the problem a little, it became
clear that whatever functionality these two commands have should be folded into "show" and
"set". It doesn't make sense to have to manipulate the data with another set of almost identical
but slightly different commands.

Also, if really needed, it's possible to use a directory metaphor for the user to navigate the data.
This might be useful also if we ever start reading in the game's database records.